requirements.txt
streamlit
torch
torchaudio
pyannote.audio
librosa==0.10.1
soundfile
openai-whisper
ffmpeg-python
